+++ packages/calcite-components/src/components/card-group/card-group.tsx
@@ -32,7 +32,7 @@ export class CardGroup extends LitElement implements InteractiveComponent {
 
   private items: Card["el"][] = [];
 
-  private slotRefEl = createRef<HTMLSlotElement>();
+  private slotRef = createRef<HTMLSlotElement>();
 
   private focusSetter = useSetFocus<this>()(this);
 
@@ -76,9 +76,7 @@ export class CardGroup extends LitElement implements InteractiveComponent {
    */
   @method()
   async setFocus(options?: FocusOptions): Promise<void> {
-    return this.focusSetter(() => {
-      return this.items[0];
-    }, options);
+    return this.focusSetter(() => this.items[0], options);
   }
 
   // #endregion
@@ -153,7 +151,7 @@ export class CardGroup extends LitElement implements InteractiveComponent {
   }
 
   private updateItemsOnSelectionModeChange(): void {
-    this.updateSlottedItems(this.slotRefEl.value);
+    this.updateSlottedItems(this.slotRef.value);
     this.updateSelectedItems();
   }
 
@@ -163,9 +161,10 @@ export class CardGroup extends LitElement implements InteractiveComponent {
   }
 
   private updateSlottedItems(target: HTMLSlotElement): void {
-    this.items = target
-      .assignedElements({ flatten: true })
-      .filter((el): el is Card["el"] => el?.matches("calcite-card"));
+    this.items =
+      target
+        ?.assignedElements({ flatten: true })
+        .filter((el): el is Card["el"] => el?.matches("calcite-card")) || [];
   }
 
   private updateSelectedItems(): void {
@@ -216,7 +215,7 @@ export class CardGroup extends LitElement implements InteractiveComponent {
     return (
       <InteractiveContainer disabled={this.disabled}>
         <div ariaLabel={this.label} class={CSS.container} role={role}>
-          <slot onSlotChange={this.updateItemsOnSlotChange} ref={this.slotRefEl} />
+          <slot onSlotChange={this.updateItemsOnSlotChange} ref={this.slotRef} />
         </div>
       </InteractiveContainer>
     );