+++ packages/calcite-components/src/components/chip/chip.tsx
@@ -2,7 +2,7 @@
 import { PropertyValues, isServer } from "lit";
 import { createRef } from "lit-html/directives/ref.js";
 import { LitElement, property, createEvent, h, method, state, JsxNode } from "@arcgis/lumina";
-import { slotChangeHasAssignedElement } from "../../utils/dom";
+import { focusElement, slotChangeHasAssignedElement } from "../../utils/dom";
 import { Appearance, Kind, Scale, SelectionMode } from "../interfaces";
 import {
   InteractiveComponent,
@@ -38,9 +38,9 @@ export class Chip extends LitElement implements InteractiveComponent {
 
   //#region Private Properties
 
-  private closeButtonEl = createRef<HTMLButtonElement>();
+  private closeButtonRef = createRef<HTMLButtonElement>();
 
-  private containerEl = createRef<HTMLDivElement>();
+  private containerRef = createRef<HTMLDivElement>();
 
   /**
    * Made into a prop for testing purposes only
@@ -146,9 +146,9 @@ export class Chip extends LitElement implements InteractiveComponent {
   async setFocus(options?: FocusOptions): Promise<void> {
     return this.focusSetter(() => {
       if (this.interactive) {
-        return this.containerEl.value;
+        return this.containerRef.value;
       } else if (this.closable) {
-        return this.closeButtonEl.value;
+        return this.closeButtonRef.value;
       }
     }, options);
   }
@@ -247,7 +247,7 @@ export class Chip extends LitElement implements InteractiveComponent {
 
   private clickHandler(): void {
     if (!this.interactive && this.closable) {
-      this.closeButtonEl.value.focus();
+      focusElement(this.closeButtonRef.value);
     }
   }
 
@@ -337,7 +337,7 @@ export class Chip extends LitElement implements InteractiveComponent {
         class={CSS.close}
         onClick={this.close}
         onKeyDown={this.closeButtonKeyDownHandler}
-        ref={this.closeButtonEl}
+        ref={this.closeButtonRef}
         tabIndex={this.disabled ? -1 : 0}
       >
         <calcite-icon icon={ICONS.close} scale={getIconScale(this.scale)} />
@@ -393,7 +393,7 @@ export class Chip extends LitElement implements InteractiveComponent {
                 (!!this.selectionMode && this.selectionMode !== "multiple" && !this.selected)),
           }}
           onClick={this.handleEmittingEvent}
-          ref={this.containerEl}
+          ref={this.containerRef}
           role={role}
           tabIndex={disableInteraction ? -1 : 0}
         >