+++ packages/calcite-components/src/components/combobox/combobox.tsx
@@ -45,7 +45,7 @@ import {
   updateHostInteraction,
 } from "../../utils/interactive";
 import { connectLabel, disconnectLabel, getLabelText, LabelableComponent } from "../../utils/label";
-import { createObserver } from "../../utils/observers";
+import { createObserver, updateRefObserver } from "../../utils/observers";
 import { toggleOpenClose, OpenCloseComponent } from "../../utils/openCloseComponent";
 import { DEBOUNCE } from "../../utils/resources";
 import { Scale, SelectionMode, Status } from "../interfaces";
@@ -101,11 +101,11 @@ export class Combobox
 
   //#region Private Properties
 
-  private closeButtonEl = createRef<HTMLButtonElement>();
+  private closeButtonRef = createRef<HTMLButtonElement>();
 
-  private selectAllComboboxItemReferenceEl = createRef<HTMLCalciteComboboxItemElement>();
+  private selectAllComboboxItemRef = createRef<HTMLCalciteComboboxItemElement>();
 
-  private allSelectedIndicatorChipEl: Chip["el"];
+  private allSelectedIndicatorChipRef = createRef<Chip["el"]>();
 
   private chipContainerEl: HTMLDivElement;
 
@@ -236,11 +236,11 @@ export class Combobox
     this.refreshSelectionDisplay();
   });
 
-  private selectedIndicatorChipEl: Chip["el"];
+  private selectedIndicatorChipRef = createRef<Chip["el"]>();
 
   private _selectedItems: HTMLCalciteComboboxItemElement["el"][] = [];
 
-  private textInput = createRef<HTMLInputElement>();
+  private textInputRef = createRef<HTMLInputElement>();
 
   transitionEl: HTMLDivElement;
 
@@ -305,12 +305,12 @@ export class Combobox
   }
 
   get keyboardNavItems(): HTMLCalciteComboboxItemElement["el"][] {
-    const { selectAllComboboxItemReferenceEl } = this;
+    const { selectAllComboboxItemRef } = this;
 
     const filteredItems = this.filteredItems.filter((item) => !item.disabled);
 
-    if (selectAllComboboxItemReferenceEl.value) {
-      return [selectAllComboboxItemReferenceEl.value, ...filteredItems];
+    if (selectAllComboboxItemRef.value) {
+      return [selectAllComboboxItemRef.value, ...filteredItems];
     }
 
     return filteredItems;
@@ -552,7 +552,7 @@ export class Combobox
     return this.focusSetter(() => {
       this.activeChipIndex = -1;
       this.activeItemIndex = -1;
-      return this.textInput.value;
+      return this.textInputRef.value;
     }, options);
   }
 
@@ -648,6 +648,10 @@ export class Combobox
 
     updateHostInteraction(this);
     this.refreshSelectionDisplay();
+  }
+
+  async load(): Promise<void> {
+    this.handleSelectionModeWarning();
   }
 
   loaded(): void {
@@ -757,9 +761,7 @@ export class Combobox
       return;
     }
     const target = event.target as HTMLCalciteComboboxItemElement["el"];
-    const isSelectAllTarget = event
-      .composedPath()
-      .includes(this.selectAllComboboxItemReferenceEl.value);
+    const isSelectAllTarget = event.composedPath().includes(this.selectAllComboboxItemRef.value);
 
     if (this.selectAllEnabled) {
       this.handleSelectAll(isSelectAllTarget);
@@ -795,7 +797,9 @@ export class Combobox
   }
 
   private clearInputValue(): void {
-    this.textInput.value.value = "";
+    if (this.textInputRef.value) {
+      this.textInputRef.value.value = "";
+    }
     this.filterText = "";
   }
 
@@ -853,7 +857,7 @@ export class Combobox
         }
         break;
       case "ArrowLeft":
-        if (this.activeChipIndex !== -1 || this.textInput.value.selectionStart === 0) {
+        if (this.activeChipIndex !== -1 || this.textInputRef.value.selectionStart === 0) {
           this.previousChip();
           event.preventDefault();
         }
@@ -895,7 +899,7 @@ export class Combobox
         }
         break;
       case " ":
-        if (!this.textInput.value.value && !event.defaultPrevented) {
+        if (!this.textInputRef.value.value && !event.defaultPrevented) {
           if (!this.open) {
             this.open = true;
             this.shiftActiveItemIndex(1);
@@ -940,7 +944,7 @@ export class Combobox
           event.preventDefault();
 
           if (this.selectAllEnabled) {
-            this.handleSelectAll(item === this.selectAllComboboxItemReferenceEl.value);
+            this.handleSelectAll(item === this.selectAllComboboxItemRef.value);
           }
         } else if (this.activeChipIndex > -1) {
           this.removeActiveChip();
@@ -1089,26 +1093,26 @@ export class Combobox
       return;
     }
 
-    if (!this.textInput.value) {
+    if (!this.textInputRef.value || !this.chipContainerEl) {
       return;
     }
 
     const {
-      allSelectedIndicatorChipEl,
+      allSelectedIndicatorChipRef,
       chipContainerEl,
       selectionDisplay,
       placeholder,
-      selectedIndicatorChipEl,
-      textInput: { value: textInput },
+      selectedIndicatorChipRef,
+      textInputRef,
     } = this;
 
     const chipContainerElGap = parseInt(getComputedStyle(chipContainerEl).gap.replace("px", ""));
     const chipContainerElWidth = getElementWidth(chipContainerEl);
-    const { fontSize, fontFamily } = getComputedStyle(textInput);
+    const { fontSize, fontFamily } = getComputedStyle(textInputRef.value);
     const inputTextWidth = getTextWidth(placeholder, `${fontSize} ${fontFamily}`);
     const inputWidth = (inputTextWidth || parseInt(calciteSize48)) + chipContainerElGap;
-    const allSelectedIndicatorChipElWidth = getElementWidth(allSelectedIndicatorChipEl);
-    const selectedIndicatorChipElWidth = getElementWidth(selectedIndicatorChipEl);
+    const allSelectedIndicatorChipElWidth = getElementWidth(allSelectedIndicatorChipRef.value);
+    const selectedIndicatorChipElWidth = getElementWidth(selectedIndicatorChipRef.value);
     const largestSelectedIndicatorChipWidth = Math.max(
       allSelectedIndicatorChipElWidth,
       selectedIndicatorChipElWidth,
@@ -1178,33 +1182,19 @@ export class Combobox
   }
 
   private setContainerEl(el: HTMLDivElement): void {
-    if (!el) {
-      return;
-    }
-
-    this.resizeObserver?.observe(el);
+    updateRefObserver(this.resizeObserver, this.listContainerEl, el);
     this.listContainerEl = el;
     this.transitionEl = el;
   }
 
   private setChipContainerEl(el: HTMLDivElement): void {
-    if (el) {
-      this.resizeObserver?.observe(el);
-    }
+    updateRefObserver(this.resizeObserver, this.chipContainerEl, el);
     this.chipContainerEl = el;
   }
 
   private setReferenceEl(el: HTMLDivElement): void {
     this.referenceEl = el;
     connectFloatingUI(this);
-  }
-
-  private setAllSelectedIndicatorChipEl(el: Chip["el"]): void {
-    this.allSelectedIndicatorChipEl = el;
-  }
-
-  private setSelectedIndicatorChipEl(el: Chip["el"]): void {
-    this.selectedIndicatorChipEl = el;
   }
 
   private setVisibleAndHiddenChips(chipEls: Chip["el"][]): void {
@@ -1304,8 +1294,8 @@ export class Combobox
     this.selectedItems = this.getSelectedItems();
     this.emitComboboxChange();
 
-    if (this.textInput.value) {
-      this.textInput.value.value = getLabel(item);
+    if (this.textInputRef.value) {
+      this.textInputRef.value.value = getLabel(item);
     }
     this.open = false;
     this.updateActiveItemIndex(-1);
@@ -1390,8 +1380,8 @@ export class Combobox
   }
 
   private resetText(): void {
-    if (this.textInput.value) {
-      this.textInput.value.value = "";
+    if (this.textInputRef.value) {
+      this.textInputRef.value.value = "";
     }
     this.filterText = "";
   }
@@ -1452,7 +1442,7 @@ export class Combobox
     const newIndex = this.activeChipIndex + 1;
     if (newIndex > last) {
       this.activeChipIndex = -1;
-      focusElement(this.textInput.value);
+      focusElement(this.textInputRef.value);
     } else {
       this.activeChipIndex = newIndex;
       this.focusChip();
@@ -1472,17 +1462,17 @@ export class Combobox
       scrollToSelected && this.selectedItems?.length
         ? this.selectedItems[0]
         : this.keyboardNavItems[this.activeItemIndex];
+    const listContainer = this.listContainerEl;
 
-    if (!item) {
+    if (!item || !listContainer) {
       return;
     }
 
     item.scrollIntoView({ block: "nearest" });
 
-    const stickyElement = this.selectAllComboboxItemReferenceEl.value;
+    const stickyElement = this.selectAllComboboxItemRef.value;
     const stickyHeight = stickyElement?.offsetHeight || 0;
 
-    const listContainer = this.listContainerEl;
     const itemRect = item.getBoundingClientRect();
     const containerRect = listContainer.getBoundingClientRect();
 
@@ -1525,7 +1515,7 @@ export class Combobox
       return;
     }
 
-    this.textInput.value?.focus();
+    this.textInputRef.value.focus();
   }
 
   private createScreenReaderItem({
@@ -1544,6 +1534,12 @@ export class Combobox
         {textContent}
       </li>
     );
+  }
+
+  private handleSelectionModeWarning(): void {
+    if (this.selectionMode === "single-persist" && this.clearDisabled) {
+      console.warn(`clearDisabled is ignored when selection-mode is set to "single-persist"`);
+    }
   }
 
   //#endregion
@@ -1594,10 +1590,10 @@ export class Combobox
 
   private renderAllSelectedIndicatorChip(): JsxNode {
     const {
+      allSelectedIndicatorChipRef,
       compactSelectionDisplay,
       scale,
       selectedVisibleChipsCount,
-      setAllSelectedIndicatorChipEl,
     } = this;
     const label = compactSelectionDisplay ? this.messages.all : this.messages.allSelected;
 
@@ -1610,7 +1606,7 @@ export class Combobox
         }}
         data-test-id="all-selected-indicator-chip"
         label={label}
-        ref={setAllSelectedIndicatorChipEl}
+        ref={allSelectedIndicatorChipRef}
         scale={scale}
         title={label}
         value=""
@@ -1628,7 +1624,7 @@ export class Combobox
       scale,
       selectedHiddenChipsCount,
       selectedVisibleChipsCount,
-      setSelectedIndicatorChipEl,
+      selectedIndicatorChipRef,
     } = this;
     let chipInvisible: boolean;
     let label: string;
@@ -1664,7 +1660,7 @@ export class Combobox
           [CSS.chipInvisible]: chipInvisible,
         }}
         label={label}
-        ref={setSelectedIndicatorChipEl}
+        ref={selectedIndicatorChipRef}
         scale={scale}
         title={label}
         value=""
@@ -1763,7 +1759,7 @@ export class Combobox
           onInput={this.inputHandler}
           placeholder={placeholder}
           readOnly={this.readOnly}
-          ref={this.textInput}
+          ref={this.textInputRef}
           required={this.required}
           role="combobox"
           tabIndex={this.activeChipIndex === -1 ? 0 : -1}
@@ -1825,7 +1821,7 @@ export class Combobox
                   id={`${this.guid}-select-all-enabled-interactive`}
                   indeterminate={this.indeterminate}
                   label={messages.selectAll}
-                  ref={this.selectAllComboboxItemReferenceEl}
+                  ref={this.selectAllComboboxItemRef}
                   scale={scale}
                   selected={this.allSelected}
                   tabIndex="-1"
@@ -1901,7 +1897,8 @@ export class Combobox
     const allSelectionDisplay = selectionDisplay === "all";
     const singleSelectionDisplay = selectionDisplay === "single";
     const fitSelectionDisplay = !singleSelectionMode && selectionDisplay === "fit";
-    const isClearable = !this.clearDisabled && this.value?.length > 0;
+    const isClearable =
+      !this.clearDisabled && this.selectionMode !== "single-persist" && !!this.value?.length;
 
     return (
       <InteractiveContainer disabled={this.disabled}>
@@ -1959,7 +1956,7 @@ export class Combobox
               disabled={this.disabled}
               key="close-button"
               label={this.messages.clear}
-              ref={this.closeButtonEl}
+              ref={this.closeButtonRef}
               scale={this.scale}
             />
           ) : null}