+++ packages/calcite-components/src/components/list/list.tsx
@@ -301,11 +301,13 @@ export class List extends LitElement implements InteractiveComponent, SortableCo
    */
   @method()
   async setFocus(options?: FocusOptions): Promise<void> {
-    return this.focusSetter(() => {
-      return this.filterEnabled
-        ? this.filterEl
-        : this.focusableItems.find((listItem) => listItem.active);
-    }, options);
+    return this.focusSetter(
+      () =>
+        this.filterEnabled
+          ? this.filterEl
+          : this.focusableItems.find((listItem) => listItem.active),
+      options,
+    );
   }
 
   /**
@@ -365,6 +367,7 @@ export class List extends LitElement implements InteractiveComponent, SortableCo
       "calciteInternalAssistiveTextChange",
       this.handleCalciteInternalAssistiveTextChange,
     );
+    this.listen("calciteListItemSortHandleBeforeOpen", this.updateListItemsDebounced);
     this.listen("calciteSortHandleReorder", this.handleSortReorder);
     this.listen("calciteSortHandleMove", this.handleSortMove);
     this.listen("calciteSortHandleAdd", this.handleSortAdd);
@@ -969,7 +972,7 @@ export class List extends LitElement implements InteractiveComponent, SortableCo
       event.preventDefault();
 
       if (currentIndex === 0 && this.filterEnabled) {
-        this.filterEl?.setFocus();
+        this.filterEl.setFocus();
         return;
       }