+++ packages/calcite-components/src/components/modal/modal.tsx
@@ -54,8 +54,6 @@ export class Modal extends LitElement implements OpenCloseComponent {
 
   //#region Private Properties
 
-  private closeButtonEl = createRef<HTMLButtonElement>();
-
   private contentId: string;
 
   private cssVarObserver: MutationObserver = createObserver("mutation", () => {
@@ -82,7 +80,7 @@ export class Modal extends LitElement implements OpenCloseComponent {
 
   private ignoreOpenChange = false;
 
-  private modalContent = createRef<HTMLDivElement>();
+  private modalContentRef = createRef<HTMLDivElement>();
 
   private mutationObserver: MutationObserver = createObserver("mutation", () =>
     this.focusTrap.updateContainerElements(),
@@ -241,12 +239,12 @@ export class Modal extends LitElement implements OpenCloseComponent {
    */
   @method()
   async scrollContent(top = 0, left = 0): Promise<void> {
-    if (this.modalContent.value) {
-      if (this.modalContent.value.scrollTo) {
-        this.modalContent.value.scrollTo({ top, left, behavior: "smooth" });
+    if (this.modalContentRef.value) {
+      if (this.modalContentRef.value.scrollTo) {
+        this.modalContentRef.value.scrollTo({ top, left, behavior: "smooth" });
       } else {
-        this.modalContent.value.scrollTop = top;
-        this.modalContent.value.scrollLeft = left;
+        this.modalContentRef.value.scrollTop = top;
+        this.modalContentRef.value.scrollLeft = left;
       }
     }
   }
@@ -260,9 +258,7 @@ export class Modal extends LitElement implements OpenCloseComponent {
    */
   @method()
   async setFocus(options?: FocusOptions): Promise<void> {
-    return this.focusSetter(() => {
-      return this.el;
-    }, options);
+    return this.focusSetter(() => this.el, options);
   }
 
   /**
@@ -517,7 +513,7 @@ export class Modal extends LitElement implements OpenCloseComponent {
               [CSS.content]: true,
               [CSS.contentNoFooter]: !this.hasFooter,
             }}
-            ref={this.modalContent}
+            ref={this.modalContentRef}
           >
             <slot name={SLOTS.content} onSlotChange={this.handleContentSlotChange} />
           </div>
@@ -567,7 +563,6 @@ export class Modal extends LitElement implements OpenCloseComponent {
         class={CSS.close}
         key="button"
         onClick={this.handleCloseClick}
-        ref={this.closeButtonEl}
         title={this.messages.close}
       >
         <calcite-icon icon={ICONS.close} scale={getIconScale(this.scale)} />