+++ packages/calcite-components/src/components/shell-panel/shell-panel.tsx
@@ -3,6 +3,7 @@ import interact from "interactjs";
 import type { Interactable, ResizeEvent } from "@interactjs/types";
 import { PropertyValues } from "lit";
 import { LitElement, property, createEvent, h, state, JsxNode } from "@arcgis/lumina";
+import { createRef } from "lit/directives/ref.js";
 import {
   getElementDir,
   getStylePixelValue,
@@ -46,7 +47,7 @@ export class ShellPanel extends LitElement {
 
   private actionBars: ActionBar["el"][] = [];
 
-  private contentEl: HTMLDivElement;
+  private contentRef = createRef<HTMLDivElement>();
 
   /**
    * Made into a prop for testing purposes only
@@ -178,7 +179,7 @@ export class ShellPanel extends LitElement {
   //#region Private Methods
 
   private getContentElDOMRect(): DOMRect {
-    return this.contentEl.getBoundingClientRect();
+    return this.contentRef.value.getBoundingClientRect();
   }
 
   private handleKeyDown(event: KeyboardEvent): void {
@@ -187,7 +188,7 @@ export class ShellPanel extends LitElement {
       position,
       layout,
       resizable,
-      contentEl,
+      contentRef,
       el,
       resizeValues: { maxBlockSize, maxInlineSize, minBlockSize, minInlineSize },
     } = this;
@@ -197,7 +198,7 @@ export class ShellPanel extends LitElement {
 
     const keys = [...arrowKeys, "Home", "End"];
 
-    if (!resizable || !contentEl || defaultPrevented || !keys.includes(key)) {
+    if (!resizable || !contentRef.value || defaultPrevented || !keys.includes(key)) {
       return;
     }
 
@@ -264,9 +265,9 @@ export class ShellPanel extends LitElement {
     type: "inlineSize" | "blockSize";
     size: number | null;
   }): void {
-    const { contentEl, resizeValues } = this;
+    const { contentRef, resizeValues } = this;
 
-    if (!contentEl) {
+    if (!contentRef.value) {
       return;
     }
 
@@ -285,7 +286,7 @@ export class ShellPanel extends LitElement {
       [type]: rounded,
     };
 
-    contentEl.style[type] = size !== null ? `${rounded}px` : null;
+    contentRef.value.style[type] = size !== null ? `${rounded}px` : null;
   }
 
   private cleanupInteractions(): void {
@@ -295,16 +296,16 @@ export class ShellPanel extends LitElement {
   private async setupInteractions(): Promise<void> {
     this.cleanupInteractions();
 
-    const { el, contentEl, resizable, position, collapsed, resizeHandleEl, layout } = this;
+    const { el, contentRef, resizable, position, collapsed, resizeHandleEl, layout } = this;
 
-    if (!contentEl || collapsed || !resizable || !resizeHandleEl) {
+    if (!contentRef.value || collapsed || !resizable || !resizeHandleEl) {
       return;
     }
 
     await this.el.componentOnReady();
 
     const { inlineSize, minInlineSize, blockSize, minBlockSize, maxInlineSize, maxBlockSize } =
-      window.getComputedStyle(contentEl);
+      window.getComputedStyle(contentRef.value);
 
     const values: ResizeValues = {
       inlineSize: getStylePixelValue(inlineSize),
@@ -319,7 +320,7 @@ export class ShellPanel extends LitElement {
 
     const rtl = getElementDir(el) === "rtl";
 
-    this.interaction = interact(contentEl, { context: el.ownerDocument }).resizable({
+    this.interaction = interact(contentRef.value, { context: el.ownerDocument }).resizable({
       edges: {
         top: position === "end" && layout === "horizontal" ? resizeHandleEl : false,
         right:
@@ -357,10 +358,6 @@ export class ShellPanel extends LitElement {
         },
       },
     });
-  }
-
-  private storeContentEl(contentEl: HTMLDivElement): void {
-    this.contentEl = contentEl;
   }
 
   private setResizeHandleEl(el: HTMLDivElement): void {
@@ -465,7 +462,7 @@ export class ShellPanel extends LitElement {
           }}
           hidden={collapsed}
           key="content"
-          ref={this.storeContentEl}
+          ref={this.contentRef}
         >
           {this.renderHeader()}
           <div class={CSS.contentBody}>