+++ packages/calcite-components/src/components/text-area/text-area.tsx
@@ -22,7 +22,7 @@ import {
 import { connectLabel, disconnectLabel, getLabelText, LabelableComponent } from "../../utils/label";
 import { slotChangeHasAssignedElement } from "../../utils/dom";
 import { NumberingSystem, numberStringFormatter } from "../../utils/locale";
-import { createObserver } from "../../utils/observers";
+import { createObserver, updateRefObserver } from "../../utils/observers";
 import {
   InteractiveComponent,
   InteractiveContainer,
@@ -78,7 +78,7 @@ export class TextArea
 
   defaultValue: TextArea["value"];
 
-  private footerEl = createRef<HTMLElement>();
+  private footerRef = createRef<HTMLElement>();
 
   private validationMessageEl: HTMLDivElement;
 
@@ -104,7 +104,7 @@ export class TextArea
       validationMessageHeight,
     } = this.getHeightAndWidthOfElements();
     if (footerWidth > 0 && footerWidth !== textAreaWidth) {
-      this.footerEl.value.style.width = `${textAreaWidth}px`;
+      this.footerRef.value.style.width = `${textAreaWidth}px`;
     }
 
     if (this.resize === "none") {
@@ -304,7 +304,7 @@ export class TextArea
   @method()
   async selectText(): Promise<void> {
     await this.componentOnReady();
-    this.textAreaEl.select();
+    this.textAreaEl?.select();
   }
 
   /**
@@ -316,9 +316,7 @@ export class TextArea
    */
   @method()
   async setFocus(options?: FocusOptions): Promise<void> {
-    return this.focusSetter(() => {
-      return this.textAreaEl;
-    }, options);
+    return this.focusSetter(() => this.textAreaEl, options);
   }
 
   //#endregion
@@ -413,11 +411,8 @@ export class TextArea
   }
 
   private setTextAreaEl(el: HTMLTextAreaElement): void {
-    if (!el) {
-      return;
-    }
+    updateRefObserver(this.resizeObserver, this.textAreaEl, el);
     this.textAreaEl = el;
-    this.resizeObserver?.observe(el);
   }
 
   private setTextAreaHeight(): void {
@@ -441,8 +436,8 @@ export class TextArea
       ? this.textAreaEl.getBoundingClientRect()
       : NO_DIMENSIONS;
     const { height: elHeight, width: elWidth } = this.el.getBoundingClientRect();
-    const { height: footerHeight, width: footerWidth } = this.footerEl.value
-      ? this.footerEl.value.getBoundingClientRect()
+    const { height: footerHeight, width: footerWidth } = this.footerRef.value
+      ? this.footerRef.value.getBoundingClientRect()
       : NO_DIMENSIONS;
 
     const { height: validationMessageHeight } = this.validationMessageEl
@@ -531,7 +526,7 @@ export class TextArea
               [CSS.readOnly]: this.readOnly,
               [CSS.hide]: !hasFooter,
             }}
-            ref={this.footerEl}
+            ref={this.footerRef}
           >
             <div
               class={{