+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.0-next.107](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.106...@esri/calcite-components-react@3.3.0-next.107) (2025-09-11)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.106](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.105...@esri/calcite-components-react@3.3.0-next.106) (2025-09-11)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.105](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.104...@esri/calcite-components-react@3.3.0-next.105) (2025-09-10)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.104](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.103...@esri/calcite-components-react@3.3.0-next.104) (2025-09-10)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.103](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.102...@esri/calcite-components-react@3.3.0-next.103) (2025-09-09)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.102](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.101...@esri/calcite-components-react@3.3.0-next.102) (2025-09-09)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.101](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.100...@esri/calcite-components-react@3.3.0-next.101) (2025-09-05)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.100](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.99...@esri/calcite-components-react@3.3.0-next.100) (2025-09-05)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.99](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.98...@esri/calcite-components-react@3.3.0-next.99) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.98](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.97...@esri/calcite-components-react@3.3.0-next.98) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.97](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.96...@esri/calcite-components-react@3.3.0-next.97) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.96](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.95...@esri/calcite-components-react@3.3.0-next.96) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.3.0-next.95](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.94...@esri/calcite-components-react@3.3.0-next.95) (2025-09-03)
 
 **Note:** Version bump only for package @esri/calcite-components-react